services/ml-service/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-multipart==0.0.20
numpy==1.26.4
scikit-learn==1.6.1
opencv-python-headless==4.10.0.84
Pillow==10.4.0
transformers==4.49.0
torch==2.5.1
